code structure commit#2
