+++ netdev/connections/__init__.py
@@ -1,0 +1,1 @@
+from .ssh import SSHConnection
