+++ netdev/connections/transport.py
@@ -1,0 +1,3 @@
+class Transport:
+    def read(self):
+        pass
