+++ netdev/vendors/devices/arista/__init__.py
@@ -1,0 +1,3 @@
+from .arista_eos import AristaEOS
+
+__all__ = ["AristaEOS"]
