+++ netdev/vendors/devices/arista/arista_eos.py
@@ -1,0 +1,7 @@
+from netdev.vendors.devices.ios_like import IOSLikeDevice
+
+
+class AristaEOS(IOSLikeDevice):
+    """Class for working with Arista EOS"""
+
+    pass
