+++ netdev/vendors/devices/aruba/aruba_aos_6.py
@@ -3,7 +3,7 @@
 import re
 
 from netdev.logger import logger
-from netdev.vendors.ios_like import IOSLikeDevice
+from netdev.vendors.devices.ios_like import IOSLikeDevice
 
 
 class ArubaAOS6(IOSLikeDevice):
