+++ netdev/vendors/devices/aruba/aruba_aos_8.py
@@ -3,7 +3,7 @@
 import re
 
 from netdev.logger import logger
-from netdev.vendors.ios_like import IOSLikeDevice
+from netdev.vendors.devices.ios_like import IOSLikeDevice
 
 
 class ArubaAOS8(IOSLikeDevice):
