+++ netdev/vendors/devices/cisco/cisco_ios.py
@@ -1,4 +1,4 @@
-from netdev.vendors.ios_like import IOSLikeDevice
+from netdev.vendors.devices.ios_like import IOSLikeDevice
 
 
 class CiscoIOS(IOSLikeDevice):
