+++ netdev/vendors/devices/cisco/cisco_iosxr.py
@@ -1,6 +1,6 @@
 from netdev.exceptions import CommitError
 from netdev.logger import logger
-from netdev.vendors.ios_like import IOSLikeDevice
+from netdev.vendors.devices.ios_like import IOSLikeDevice
 
 
 class CiscoIOSXR(IOSLikeDevice):
