+++ netdev/vendors/devices/comware_like.py
@@ -7,7 +7,7 @@ Connection Method are based upon AsyncSSH and should be running in asyncio loop
 import re
 
 from netdev.logger import logger
-from netdev.vendors.base import BaseDevice
+from netdev.vendors.devices.base import BaseDevice
 
 
 class ComwareLikeDevice(BaseDevice):
