+++ netdev/vendors/devices/hp/hp_comware.py
@@ -1,4 +1,4 @@
-from netdev.vendors.comware_like import ComwareLikeDevice
+from netdev.vendors.devices.comware_like import ComwareLikeDevice
 
 
 class HPComware(ComwareLikeDevice):
