+++ netdev/vendors/terminal_modes/__init__.py
@@ -1,0 +1,1 @@
+from .common import TerminalMode
