--- netdev/vendors/transport.py
@@ -1,3 +1,0 @@
-class Transport:
-    def __init__(self, stdin, stdout, stderr):
-        pass
